chore(fi): factor out version parser into module
This makes it clear that they are the same thing and prevents drifting.
